+++ src/app/products/components/product/product.component.ts
@@ -1,14 +1,17 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { ProductModelInterface } from '../../models/product-model-interface';
 
 @Component({
   selector: 'app-product',
   templateUrl: './product.component.html',
-  styleUrls: ['./product.component.css']
+  styleUrls: ['./product.component.css'],
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class ProductComponent implements OnInit {
 
   @Input() product: ProductModelInterface | undefined;
+
+  @Output() buyClicked = new EventEmitter<number>();
 
   constructor() { }
 
@@ -16,6 +19,6 @@ export class ProductComponent implements OnInit {
   }
 
   onBuy(): void {
-    console.log('Product added');
+    this.buyClicked.emit(this.product?.id);
   }
 }