+++ src/app/products/models/product-model-interface.ts
@@ -1,6 +1,7 @@
 import { CategoryEnum } from './category-enum.enum';
 
 export interface ProductModelInterface {
+  id: number;
   name: string;
   description: string;
   price: number;