+++ src/app/products/services/products.service.ts
@@ -10,8 +10,8 @@ export class ProductsService {
   constructor() { }
 
   getProducts(): ProductModelInterface[] {
-    return [{name: 'A product', price: 1, isAvailable: true, category: CategoryEnum.Category1, description: 'Test Description 1'},
-      {name: 'B product', price: 2, isAvailable: true, category: CategoryEnum.Category2, description: 'Test Description 2'},
-      {name: 'C product', price: 3, isAvailable: false, category: CategoryEnum.Category2, description: 'Test Description 3'}];
+    return [{id: 1, name: 'A product', price: 100, isAvailable: true, category: CategoryEnum.Category1, description: 'Test Description 1'},
+      {id: 2, name: 'B product', price: 90, isAvailable: true, category: CategoryEnum.Category2, description: 'Test Description 2'},
+      {id: 3, name: 'C product', price: 125, isAvailable: false, category: CategoryEnum.Category2, description: 'Test Description 3'}];
   }
 }